refactor: 패키지 정리, 변수 순서 정리
