+++ src/main/java/lotto/model/LottoChecker.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import lotto.Rank;
+import lotto.util.Rank;
 
 public class LottoChecker {
