+++ src/main/java/lotto/util/Constant.java
@@ -2,13 +2,15 @@ package lotto.util;
 
 public final class Constant {
 
-    private Constant() {
-
-    }
+    private Constant() {}
 
     public static final int ZERO = 0;
+
     public static final int MIN_LOTTO_NUM = 1;
+
     public static final int MAX_LOTTO_NUM = 45;
+
     public static final int LOTTO_NUM_COUNT = 6;
+
     public static final int LOTTO_PRICE = 1000;
 }
